# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for TutorCall
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomBody {
  hostName: string;
}

export interface Room {
  id: string;
  hostName: string;
  createdAt: string;
  participantCount: number;
}

export interface Participant {
  id: string;
  name: string;
  isHost: boolean;
  isMuted: boolean;
  isVideoOff: boolean;
  joinedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetRoomParticipants200 = {
  participants: Participant[];
};