# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for TutorCall
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Room management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room info
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomId}/participants:
    get:
      operationId: getRoomParticipants
      tags: [rooms]
      summary: Get room participants
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant list
          content:
            application/json:
              schema:
                type: object
                properties:
                  participants:
                    type: array
                    items:
                      $ref: "#/components/schemas/Participant"
                required:
                  - participants
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateRoomBody:
      type: object
      properties:
        hostName:
          type: string
      required:
        - hostName
    Room:
      type: object
      properties:
        id:
          type: string
        hostName:
          type: string
        createdAt:
          type: string
        participantCount:
          type: integer
      required:
        - id
        - hostName
        - createdAt
        - participantCount
    Participant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        isHost:
          type: boolean
        isMuted:
          type: boolean
        isVideoOff:
          type: boolean
        joinedAt:
          type: string
      required:
        - id
        - name
        - isHost
        - isMuted
        - isVideoOff
        - joinedAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error