# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for TutorCall
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new room
 */
export const CreateRoomBody = zod.object({
  hostName: zod.string(),
});

/**
 * @summary Get room info
 */
export const GetRoomParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  id: zod.string(),
  hostName: zod.string(),
  createdAt: zod.string(),
  participantCount: zod.number(),
});

/**
 * @summary Get room participants
 */
export const GetRoomParticipantsParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetRoomParticipantsResponse = zod.object({
  participants: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      isHost: zod.boolean(),
      isMuted: zod.boolean(),
      isVideoOff: zod.boolean(),
      joinedAt: zod.string(),
    }),
  ),
});